Float could be double
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/pocketsphinx-android@12535 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/edu/cmu/pocketsphinx/SpeechRecognizerSetup.java
@@ -103,7 +103,7 @@ public class SpeechRecognizerSetup {
         return this;
     }
 
-    public SpeechRecognizerSetup setFloat(String key, float value) {
+    public SpeechRecognizerSetup setFloat(String key, double value) {
         config.setFloat(key, value);
         return this;
     }
